+++ hydra-node/src/Hydra/API/HTTPServer.hs
@@ -5,6 +5,7 @@ module Hydra.API.HTTPServer where
 import Hydra.Prelude
 
 import Cardano.Ledger.Core (PParams)
+import Control.Concurrent.STM (TChan, dupTChan, readTChan)
 import Data.Aeson (KeyValue ((.=)), object, withObject, (.:))
 import Data.Aeson qualified as Aeson
 import Data.ByteString.Lazy qualified as LBS
@@ -12,37 +13,20 @@ import Data.ByteString.Short ()
 import Data.Text (pack)
 import Hydra.API.APIServerLog (APIServerLog (..), Method (..), PathInfo (..))
 import Hydra.API.ClientInput (ClientInput (..))
-import Hydra.API.ServerOutput (CommitInfo (..), getConfirmedSnapshot, getSeenSnapshot, getSnapshotUtxo)
-import Hydra.Cardano.Api (
-  LedgerEra,
-  Tx,
- )
+import Hydra.API.ServerOutput (ClientMessage, CommitInfo (..), ServerOutput (..), TimedServerOutput (..), getConfirmedSnapshot, getSeenSnapshot, getSnapshotUtxo)
+import Hydra.Cardano.Api (LedgerEra, Tx)
 import Hydra.Chain (Chain (..), PostTxError (..), draftCommitTx)
-import Hydra.Chain.ChainState (
-  IsChainState,
- )
+import Hydra.Chain.ChainState (IsChainState)
 import Hydra.Chain.Direct.State ()
-import Hydra.HeadLogic.State (
-  HeadState (..),
- )
+import Hydra.HeadLogic.State (HeadState (..))
+import Hydra.Ledger (ValidationError (..))
 import Hydra.Logging (Tracer, traceWith)
+import Hydra.Node.ApiTransactionTimeout (ApiTransactionTimeout (..))
 import Hydra.Node.DepositPeriod (toNominalDiffTime)
 import Hydra.Node.Environment (Environment (..))
-import Hydra.Tx (
-  CommitBlueprintTx (..),
-  ConfirmedSnapshot,
-  IsTx (..),
-  UTxOType,
- )
-import Network.HTTP.Types (ResponseHeaders, hContentType, status200, status400, status404, status500)
-import Network.Wai (
-  Application,
-  Request (pathInfo, requestMethod),
-  Response,
-  consumeRequestBodyStrict,
-  rawPathInfo,
-  responseLBS,
- )
+import Hydra.Tx (CommitBlueprintTx (..), ConfirmedSnapshot, IsTx (..), Snapshot (..), UTxOType)
+import Network.HTTP.Types (ResponseHeaders, hContentType, status200, status202, status400, status404, status500)
+import Network.Wai (Application, Request (pathInfo, requestMethod), Response, consumeRequestBodyStrict, rawPathInfo, responseLBS)
 
 newtype DraftCommitTxResponse tx = DraftCommitTxResponse
   { commitTx :: tx
@@ -141,6 +125,49 @@ instance (Arbitrary tx, Arbitrary (UTxOType tx), IsTx tx) => Arbitrary (SideLoad
   shrink = \case
     SideLoadSnapshotRequest snapshot -> SideLoadSnapshotRequest <$> shrink snapshot
 
+-- | Request to submit a transaction to the head
+newtype SubmitL2TxRequest tx = SubmitL2TxRequest
+  { submitL2Tx :: tx
+  }
+  deriving newtype (Eq, Show, Arbitrary)
+  deriving newtype (ToJSON, FromJSON)
+
+-- | Response for transaction submission
+data SubmitL2TxResponse
+  = -- | Transaction was included in a confirmed snapshot
+    SubmitTxConfirmed Integer
+  | -- | Transaction was rejected due to validation errors
+    SubmitTxInvalidResponse Text
+  | -- | Transaction was accepted but not yet confirmed
+    SubmitTxSubmitted
+  deriving stock (Eq, Show, Generic)
+
+instance ToJSON SubmitL2TxResponse where
+  toJSON = \case
+    SubmitTxConfirmed snapshotNumber ->
+      object
+        [ "tag" .= Aeson.String "SubmitTxConfirmed"
+        , "snapshotNumber" .= snapshotNumber
+        ]
+    SubmitTxInvalidResponse validationError ->
+      object
+        [ "tag" .= Aeson.String "SubmitTxInvalid"
+        , "validationError" .= validationError
+        ]
+    SubmitTxSubmitted -> object ["tag" .= Aeson.String "SubmitTxSubmitted"]
+
+instance FromJSON SubmitL2TxResponse where
+  parseJSON = withObject "SubmitTxResponse" $ \o -> do
+    tag <- o .: "tag"
+    case tag :: Text of
+      "SubmitTxConfirmed" -> SubmitTxConfirmed <$> o .: "snapshotNumber"
+      "SubmitTxInvalid" -> SubmitTxInvalidResponse <$> o .: "validationError"
+      "SubmitTxSubmitted" -> pure SubmitTxSubmitted
+      _ -> fail "Expected tag to be SubmitTxConfirmed, SubmitTxInvalid, or SubmitTxSubmitted"
+
+instance Arbitrary SubmitL2TxResponse where
+  arbitrary = genericArbitrary
+
 jsonContent :: ResponseHeaders
 jsonContent = [(hContentType, "application/json")]
 
@@ -160,8 +187,12 @@ httpApp ::
   IO [TxIdType tx] ->
   -- | Callback to yield a 'ClientInput' to the main event loop.
   (ClientInput tx -> IO ()) ->
+  -- | Timeout for transaction submission
+  ApiTransactionTimeout ->
+  -- | Channel to listen for events
+  TChan (Either (TimedServerOutput tx) (ClientMessage tx)) ->
   Application
-httpApp tracer directChain env pparams getHeadState getCommitInfo getPendingDeposits putClientInput request respond = do
+httpApp tracer directChain env pparams getHeadState getCommitInfo getPendingDeposits putClientInput apiTransactionTimeout responseChannel request respond = do
   traceWith tracer $
     APIHTTPRequestReceived
       { method = Method $ requestMethod request
@@ -206,6 +237,10 @@ httpApp tracer directChain env pparams getHeadState getCommitInfo getPendingDepo
     ("POST", ["cardano-transaction"]) ->
       consumeRequestBodyStrict request
         >>= handleSubmitUserTx directChain
+        >>= respond
+    ("POST", ["transaction"]) ->
+      consumeRequestBodyStrict request
+        >>= handleSubmitL2Tx putClientInput apiTransactionTimeout responseChannel
         >>= respond
     _ ->
       respond $ responseLBS status400 jsonContent . Aeson.encode $ Aeson.String "Resource not found"
@@ -337,6 +372,74 @@ handleSideLoadSnapshot putClientInput body = do
     Right SideLoadSnapshotRequest{snapshot} -> do
       putClientInput $ SideLoadSnapshot snapshot
       pure $ responseLBS status200 jsonContent (Aeson.encode $ Aeson.String "OK")
+
+-- | Handle request to submit a transaction to the head.
+handleSubmitL2Tx ::
+  forall tx.
+  IsChainState tx =>
+  (ClientInput tx -> IO ()) ->
+  ApiTransactionTimeout ->
+  TChan (Either (TimedServerOutput tx) (ClientMessage tx)) ->
+  LBS.ByteString ->
+  IO Response
+handleSubmitL2Tx putClientInput apiTransactionTimeout responseChannel body = do
+  case Aeson.eitherDecode' @(SubmitL2TxRequest tx) body of
+    Left err ->
+      pure $ responseLBS status400 jsonContent (Aeson.encode $ Aeson.String $ pack err)
+    Right SubmitL2TxRequest{submitL2Tx} -> do
+      -- Duplicate the channel to avoid consuming messages from other consumers.
+      dupChannel <- atomically $ dupTChan responseChannel
+
+      -- Submit the transaction to the head
+      putClientInput (NewTx submitL2Tx)
+
+      let txid = txId submitL2Tx
+      result <-
+        timeout
+          (realToFrac (apiTransactionTimeoutNominalDiffTime apiTransactionTimeout))
+          (waitForTransactionResult dupChannel txid)
+
+      case result of
+        Just (SubmitTxConfirmed snapshotNumber) ->
+          pure $ responseLBS status200 jsonContent (Aeson.encode $ SubmitTxConfirmed snapshotNumber)
+        Just (SubmitTxInvalidResponse validationError) ->
+          pure $ responseLBS status400 jsonContent (Aeson.encode $ SubmitTxInvalidResponse validationError)
+        Just SubmitTxSubmitted ->
+          pure $ responseLBS status202 jsonContent (Aeson.encode SubmitTxSubmitted)
+        Nothing ->
+          -- Timeout occurred - return 202 Accepted with timeout info
+          pure $
+            responseLBS
+              status202
+              jsonContent
+              ( Aeson.encode $
+                  object
+                    [ "tag" .= Aeson.String "SubmitTxSubmitted"
+                    , "timeout" .= Aeson.String ("Transaction submission timed out after " <> pack (show apiTransactionTimeout) <> " seconds")
+                    ]
+              )
+ where
+  --  Wait for transaction result by listening to events
+  waitForTransactionResult :: TChan (Either (TimedServerOutput tx) (ClientMessage tx)) -> TxIdType tx -> IO SubmitL2TxResponse
+  waitForTransactionResult dupChannel txid = go
+   where
+    go = do
+      event <- atomically $ readTChan dupChannel
+      case event of
+        Left (TimedServerOutput{output}) -> case output of
+          TxValid{transactionId}
+            | transactionId == txid ->
+                pure SubmitTxSubmitted
+          TxInvalid{transaction, validationError = ValidationError reason}
+            | txId transaction == txid ->
+                pure $ SubmitTxInvalidResponse reason
+          SnapshotConfirmed{snapshot} ->
+            -- Check if the transaction is in the confirmed snapshot
+            if txid `elem` map txId (confirmed snapshot)
+              then pure $ SubmitTxConfirmed (fromIntegral $ number snapshot)
+              else go
+          _ -> go
+        Right _ -> go
 
 badRequest :: IsChainState tx => PostTxError tx -> Response
 badRequest = responseLBS status400 jsonContent . Aeson.encode . toJSON