+++ hydra-node/test/Hydra/API/HTTPServerSpec.hs
@@ -4,6 +4,7 @@ import Hydra.Prelude hiding (get)
 import Test.Hydra.Prelude
 
 import Cardano.Api.UTxO qualified as UTxO
+import Control.Concurrent.STM (newTChanIO, writeTChan)
 import Control.Lens ((^?))
 import Data.Aeson (Result (Error, Success), eitherDecode, encode, fromJSON)
 import Data.Aeson qualified as Aeson
@@ -13,11 +14,13 @@ import Hydra.API.HTTPServer (
   DraftCommitTxRequest (..),
   DraftCommitTxResponse (..),
   SideLoadSnapshotRequest (..),
+  SubmitL2TxRequest (..),
+  SubmitL2TxResponse (..),
   SubmitTxRequest (..),
   TransactionSubmitted,
   httpApp,
  )
-import Hydra.API.ServerOutput (CommitInfo (CannotCommit, NormalCommit), getConfirmedSnapshot, getSeenSnapshot, getSnapshotUtxo)
+import Hydra.API.ServerOutput (CommitInfo (CannotCommit, NormalCommit), ServerOutput (..), TimedServerOutput (..), getConfirmedSnapshot, getSeenSnapshot, getSnapshotUtxo)
 import Hydra.API.ServerSpec (dummyChainHandle)
 import Hydra.Cardano.Api (
   mkTxOutDatumInline,
@@ -30,8 +33,9 @@ import Hydra.Chain.Direct.Handlers (rejectLowDeposits)
 import Hydra.HeadLogic.State (ClosedState (..), HeadState (..), SeenSnapshot (..))
 import Hydra.HeadLogicSpec (inIdleState)
 import Hydra.JSONSchema (SchemaSelector, prop_validateJSONSchema, validateJSON, withJsonSpecifications)
+import Hydra.Ledger (ValidationError (..))
 import Hydra.Ledger.Cardano (Tx)
-import Hydra.Ledger.Simple (SimpleTx)
+import Hydra.Ledger.Simple (SimpleTx (..))
 import Hydra.Logging (nullTracer)
 import Hydra.Tx (ConfirmedSnapshot (..))
 import Hydra.Tx.IsTx (UTxOType)
@@ -63,6 +67,8 @@ spec = do
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized TransactionSubmitted))
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized (SideLoadSnapshotRequest Tx)))
     roundtripAndGoldenSpecs (Proxy @(ReasonablySized (HeadState Tx)))
+    roundtripAndGoldenSpecs (Proxy @(ReasonablySized SubmitL2TxResponse))
+    roundtripAndGoldenSpecs (Proxy @(ReasonablySized (SubmitL2TxRequest Tx)))
 
     prop "Validate /commit publish api schema" $
       prop_validateJSONSchema @(DraftCommitTxRequest Tx) "api.json" $
@@ -168,6 +174,20 @@ spec = do
           . key "subscribe"
           . key "message"
 
+    prop "Validate /transaction publish api schema" $
+      prop_validateJSONSchema @(SubmitL2TxRequest Tx) "api.json" $
+        key "channels"
+          . key "/transaction"
+          . key "publish"
+          . key "message"
+          . key "payload"
+
+    prop "Validate /transaction subscribe api schema" $
+      prop_validateJSONSchema @SubmitL2TxResponse "api.json" $
+        key "components"
+          . key "schemas"
+          . key "SubmitL2TxResponse"
+
     apiServerSpec
     describe "SubmitTxRequest accepted tx formats" $ do
       prop "accepts json encoded transaction" $
@@ -191,6 +211,7 @@ apiServerSpec = do
         putClientInput = const (pure ())
         getHeadState = pure inIdleState
     describe "GET /protocol-parameters" $ do
+      responseChannel <- runIO newTChanIO
       with
         ( return $
             httpApp @SimpleTx
@@ -202,6 +223,8 @@ apiServerSpec = do
               cantCommit
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
         )
         $ do
           it "matches schema" $
@@ -220,6 +243,7 @@ apiServerSpec = do
                 }
 
     describe "GET /head" $ do
+      responseChannel <- runIO newTChanIO
       prop "responds correctly" $ \headState -> do
         withApplication
           ( httpApp @SimpleTx
@@ -231,10 +255,13 @@ apiServerSpec = do
               cantCommit
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
           )
           $ do
             get "/head"
               `shouldRespondWith` 200{matchBody = matchJSON headState}
+      responseChannelSimpleTx <- runIO newTChanIO
       prop "ok response matches schema" $ \headState -> do
         let isIdle = case headState of
               Idle{} -> True
@@ -265,6 +292,8 @@ apiServerSpec = do
                   cantCommit
                   getPendingDeposits
                   putClientInput
+                  300
+                  responseChannelSimpleTx
               )
               $ do
                 get "/head"
@@ -275,6 +304,7 @@ apiServerSpec = do
                           (key "channels" . key "/head" . key "subscribe" . key "message")
                     }
     describe "GET /snapshot/last-seen" $ do
+      responseChannel <- runIO newTChanIO
       prop "responds correctly" $ \headState -> do
         let seenSnapshot :: SeenSnapshot SimpleTx = getSeenSnapshot headState
         withApplication
@@ -287,11 +317,14 @@ apiServerSpec = do
               cantCommit
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
           )
           $ do
             get "/snapshot/last-seen"
               `shouldRespondWith` 200{matchBody = matchJSON seenSnapshot}
     describe "GET /snapshot" $ do
+      responseChannel <- runIO newTChanIO
       prop "responds correctly" $ \headState -> do
         let confirmedSnapshot :: Maybe (ConfirmedSnapshot SimpleTx) = getConfirmedSnapshot headState
         withApplication
@@ -304,13 +337,16 @@ apiServerSpec = do
               cantCommit
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
           )
           $ do
             get "/snapshot"
               `shouldRespondWith` case confirmedSnapshot of
                 Nothing -> 404
                 Just confirmedSn -> 200{matchBody = matchJSON confirmedSn}
-      prop "ok response matches schema" $ \(closedState :: ClosedState tx) ->
+      responseChannelSimpleTx <- runIO newTChanIO
+      prop "ok response matches schema" $ \(closedState :: ClosedState tx) -> do
         withMaxSuccess 4
           . withJsonSpecifications
           $ \schemaDir -> do
@@ -324,6 +360,8 @@ apiServerSpec = do
                   cantCommit
                   getPendingDeposits
                   putClientInput
+                  300
+                  responseChannelSimpleTx
               )
               $ do
                 get "/snapshot"
@@ -335,6 +373,7 @@ apiServerSpec = do
                     }
 
     describe "POST /snapshot" $ do
+      responseChannel <- runIO newTChanIO
       prop "responds on valid requests" $ \(request :: SideLoadSnapshotRequest Tx, headState) -> do
         withMaxSuccess 10
           . withApplication
@@ -347,12 +386,15 @@ apiServerSpec = do
                 cantCommit
                 getPendingDeposits
                 putClientInput
+                300
+                responseChannel
             )
           $ do
             post "/snapshot" (Aeson.encode request)
             `shouldRespondWith` 200
 
     describe "GET /snapshot/utxo" $ do
+      responseChannel <- runIO newTChanIO
       prop "responds correctly" $ \headState -> do
         let utxo :: Maybe (UTxOType SimpleTx) = getSnapshotUtxo headState
         withApplication
@@ -365,12 +407,15 @@ apiServerSpec = do
               cantCommit
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
           )
           $ do
             get "/snapshot/utxo"
               `shouldRespondWith` case utxo of
                 Nothing -> 404
                 Just u -> 200{matchBody = matchJSON u}
+      responseChannelSimpleTx <- runIO newTChanIO
       prop "ok response matches schema" $ \headState -> do
         let mUTxO = getSnapshotUtxo headState
             utxo :: UTxOType Tx = fromMaybe mempty mUTxO
@@ -389,6 +434,8 @@ apiServerSpec = do
                   cantCommit
                   getPendingDeposits
                   putClientInput
+                  300
+                  responseChannelSimpleTx
               )
               $ do
                 get "/snapshot/utxo"
@@ -425,6 +472,8 @@ apiServerSpec = do
                 cantCommit
                 getPendingDeposits
                 putClientInput
+                300
+                responseChannelSimpleTx
             )
             $ do
               get "/snapshot/utxo"
@@ -445,6 +494,7 @@ apiServerSpec = do
               }
       let initialHeadState = Initial (generateWith arbitrary 42)
       let openHeadState = Open (generateWith arbitrary 42)
+      responseChannel <- runIO newTChanIO
       prop "responds on valid requests" $ \(request :: DraftCommitTxRequest Tx) ->
         withApplication
           ( httpApp
@@ -456,6 +506,8 @@ apiServerSpec = do
               getHeadId
               getPendingDeposits
               putClientInput
+              300
+              responseChannel
           )
           $ do
             post "/commit" (Aeson.encode request)
@@ -497,6 +549,8 @@ apiServerSpec = do
                 getHeadId
                 getPendingDeposits
                 putClientInput
+                300
+                responseChannel
             )
           $ do
             post "/commit" (Aeson.encode (request :: DraftCommitTxRequest Tx))
@@ -506,6 +560,95 @@ apiServerSpec = do
                 CannotFindOwnInitial{} -> 400
                 DepositTooLow{} -> 400
                 _ -> 500
+
+    describe "POST /transaction" $ do
+      let mkReq tx = encode $ SubmitL2TxRequest tx
+          testTx = SimpleTx 42 mempty mempty
+          testHeadId = generateWith arbitrary 42
+      now <- runIO getCurrentTime
+
+      prop "returns 202 Accepted on timeout" $ do
+        responseChannel <- newTChanIO
+        withApplication
+          ( httpApp @SimpleTx
+              nullTracer
+              dummyChainHandle
+              testEnvironment
+              defaultPParams
+              (pure inIdleState)
+              (pure CannotCommit)
+              (pure [])
+              (const $ pure ())
+              0
+              responseChannel
+          )
+          $ do
+            post "/transaction" (mkReq testTx) `shouldRespondWith` 202
+
+      prop "returns 200 OK on confirmed snapshot" $ do
+        responseChannel <- newTChanIO
+        let snapshot =
+              Snapshot
+                { headId = testHeadId
+                , version = 1
+                , number = 7
+                , confirmed = [testTx]
+                , utxo = mempty
+                , utxoToCommit = mempty
+                , utxoToDecommit = mempty
+                }
+            event =
+              TimedServerOutput
+                { output = SnapshotConfirmed{snapshot = snapshot, signatures = mempty, headId = testHeadId}
+                , seq = 0
+                , time = now
+                }
+        withApplication
+          ( httpApp @SimpleTx
+              nullTracer
+              dummyChainHandle
+              testEnvironment
+              defaultPParams
+              (pure inIdleState)
+              (pure CannotCommit)
+              (pure [])
+              (const $ atomically $ writeTChan responseChannel (Left event))
+              10
+              responseChannel
+          )
+          $ do
+            post "/transaction" (mkReq testTx) `shouldRespondWith` 200
+
+      prop "returns 400 Bad Request on invalid tx" $ do
+        responseChannel <- newTChanIO
+        let validationError = ValidationError "some error"
+            event =
+              TimedServerOutput
+                { output =
+                    TxInvalid
+                      { headId = testHeadId
+                      , utxo = mempty
+                      , transaction = testTx
+                      , validationError = validationError
+                      }
+                , seq = 0
+                , time = now
+                }
+        withApplication
+          ( httpApp @SimpleTx
+              nullTracer
+              dummyChainHandle
+              testEnvironment
+              defaultPParams
+              (pure inIdleState)
+              (pure CannotCommit)
+              (pure [])
+              (const $ atomically $ writeTChan responseChannel (Left event))
+              10
+              responseChannel
+          )
+          $ do
+            post "/transaction" (mkReq testTx) `shouldRespondWith` 400
 
 -- * Helpers